% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFAM2PDB.R
\name{choose_PDBs_for_a_pfamid}
\alias{choose_PDBs_for_a_pfamid}
\title{Download and annotate the PDB files associated with the PFAM domain sequences in the sth file}
\usage{
choose_PDBs_for_a_pfamid(
  pfamid = "PF00075",
  pfamSthFile = NA,
  max2take = 12,
  resolution.th = 3,
  experimentalTech = "X-RAY DIFFRACTION",
  pfamseqResource = "rp75",
  pdb.local.db,
  sth.local.db
)
}
\arguments{
\item{pfamid}{A Pfam id string.}

\item{pfamSthFile}{A pfam sth file for pfamid.}

\item{max2take}{The maximal number of PDB files to download with the given resolution (default = 3.0).}

\item{resolution.th}{Minimal resoultion required for a solved structure to be considered.}

\item{experimentalTech}{Experimental method required for the structre (default is X-RAY).}

\item{pfamseqResource}{A sequence redundancy level as inidcated in Pfam - relevant when downloading the sth file.}

\item{pdb.local.db}{A directory path to a local pdb database or temporary directory to save downloaded sth files.
Can be used to accomulate files and avoid repeated downloads.}

\item{sth.local.db}{A directory path to a local pfam database or temporary directory to save downloaded sth files.
Can be used to accomulate files and avoid repeated downloads.}
}
\value{
A map of all known PDB residues to the msa file by PDB chain. TODO: add description
}
\description{
Download and annotate the PDB files associated with the PFAM domain sequences in the sth file
}
\examples{
#In this example we download all relevant relevant PDB files pfam domain PF00075.
#We limit the download to a total of 12 pdb instances with a resolution below 3.0 Ang.
#If any of the pdb files was already downloaded and located in the pdb.local.db,
#then it will be used to avoid
#unnecessary download.
tmp <- choose_PDBs_for_a_pfamid(pfamid = "PF00075", pfamSthFile = "/tmp/PF00075_rp75.sth",
                                pdb.local.db = "/tmp/", sth.local.db = "/tmp/", max2take = 3)
}
