% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MAPDB.R
\docType{data}
\name{hmmer_pdb_all}
\alias{hmmer_pdb_all}
\title{Mapping between Pfam domains and PDB chains.}
\format{A data frame with 130678 rows and 8 variables:
\describe{
  \item{PDB_ID}{4 letters PDB code}
  \item{CHAIN_ID}{PDB chain id}
  \item{PdbResNumStart}{PDB residue number where the pfam domain starts}
  \item{PdbResNumEnd}{PDB residue number where the pfam domain ends}
  \item{PFAM_ACC}{Pfam id}
  ...
}}
\source{
\url{http://www.rcsb.org/pdb/rest/hmmer?file=hmmer_pdb_all.txt}
}
\usage{
hmmer_pdb_all
}
\description{
A dataset containing the mapping between Pfam domains (based on HMMER) and PDB for more than
130,000 PDB chains.
}
\keyword{datasets}

