% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPDB.R
\docType{package}
\name{MAPDB}
\alias{MAPDB}
\title{MAPDB: A package for mapping Pfam multiple sequence alignment (MSA in sth format) or Uniprot (MSA in fasta format)
to PDB residue coordinates.}
\description{
MAPDB provides an easy way to map all known PDB chains to an STH or fasta MSA file.
}
\details{
using this package requires an interent connection. If more then several examples are required,
you should consider downloading first PDB and STH databases.

This package can be used, for example, to calculate the accuracy of contact predictions based on MSA data.
If you use this package please cite:

Etai Jacob, Ron Unger and Amnon Horovitz (2015). Codon-level information improves predictions of inter-residue contacts
in proteins by correlated mutation analysis.
eLife 2015;10.7554/eLife.08932 URL http://dx.doi.org/10.7554/eLife.08932.
It has two main functions:

\itemize{
\item Map all known PDB residues' coordinates to a Pfam STH file
\item Map all known PDB residues' coordinates to a fasta file (partially supported at this version)
}

To learn more about MAPDB, start with the vignettes:
\code{browseVignettes(package = "MAPDB")}
}
