% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MAPDB.R
\docType{data}
\name{pdb_chain_uniprot}
\alias{pdb_chain_uniprot}
\title{Mapping between PDB chains and uniprot entries.}
\format{A data frame with 294960 rows and 9 variables:
\describe{
  \item{PDB}{4 letters PDB code}
  \item{CHAIN}{PDB chain id}
  \item{RES_BEG}{Uniprot position where PDB chain starts}
  \item{RES_END}{Uniprot position where PDB chain ends}
  \item{PDB_BEG}{PDB residue number where the uniprot sequence starts}
  \item{PDB_END}{PDB residue number where the uniprot sequence ends}
  ...
}}
\source{
\url{ftp://ftp.ebi.ac.uk/pub/databases/msd/sifts/text/pdb_chain_uniprot.lst}
}
\usage{
pdb_chain_uniprot
}
\description{
A dataset containing the mapping between PDB chains and Uniprot entries.
}
\keyword{datasets}

