% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MAPDB.R
\name{PDBs2Pfam}
\alias{PDBs2Pfam}
\title{Map all known PDB residues' coordinates to an Pfam STH file.}
\usage{
PDBs2Pfam(pfamid, pfamSthFile = NA, sth.local.db = "data-raw/STH/",
  pfamseqResource = "rp75", ...)
}
\arguments{
\item{pfamid}{A Pfam id string}

\item{pfamSthFile}{A pfam sth file for pfamid.}

\item{sth.local.db}{A path to a local pfam database or temporary directory to save downloaded sth files.}

\item{pfamseqResource}{A sequence redundancy level as inidcated in Pfam - relevant when downloading the sth file.}
}
\value{
A map of all known PDB residues to the msa file by PDB chain.
}
\description{
Map all known PDB residues' coordinates to an Pfam STH file.
}
\examples{
myPfamMap <- PDBs2Pfam("PF00075")
}

