% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPDB.R
\docType{data}
\name{pdbUniprot}
\alias{pdbUniprot}
\title{Mapping between PDB chains and swissprot accs and ids.}
\format{A data frame with 229185 rows and 15 variables:
\describe{
  \item{SP_PRIMARY}{id}
  \item{PDB}{4 letters PDB code}
  \item{CHAIN}{PDB chain id}
  \item{RES_BEG}{Uniprot position where PDB chain starts}
  \item{RES_END}{Uniprot position where PDB chain ends}
  \item{PDB_BEG}{PDB residue number where the uniprot sequence starts}
  \item{PDB_END}{PDB residue number where the uniprot sequence ends}
  \item{SP_BEG}{start position}
  \item{SP_END}{end position}
  \item{Entry.name}{Entry name}
}}
\source{
\url{http://www.uniprot.org/uniprot}
}
\usage{
pdbUniprot
}
\description{
A processed dataset of almost 230,000 rows containing the mapping between PDB chains and swissprot entries.
The list is based on merging pdb_chain_uniprot.lst file from EBI and Uniprot. For more inforamtion see pdbUniprot.R in this
Package.
}
\keyword{datasets}
